requirements.txt
streamlit>=1.22.0
pyautogen>=0.2.19
ipython
huggingface_hub
matplotlib
finnhub-python
yfinance
mplfinance
backtrader
sec_api
numpy
pandas
pyPDF2
reportlab
pyautogen[retrievechat]
praw
tushare
pandas_datareader
marker-pdf
langchain==0.1.20
nltk==3.8.1
scikit_learn==1.5.0
tenacity==8.3.0
tqdm==4.66.1
typing_extensions==4.9.0
unstructured==0.8.1